Moved commands cd and ls into commands package
