+++ src/main/java/intecmd/commands/ChangeDirectory.java
@@ -1,4 +1,6 @@
-package intecmd;
+package intecmd.commands;
+
+import intecmd.CurrentDirectory;
 
 import java.io.File;
 import java.util.ArrayList;
